+++ src/lib/vault.js
@@ -1,6 +1,7 @@
 const CryptoJS = require('crypto-js');
 const { KeyringController } = require('@getsafle/vault-eth-controller');
 const BitcoinKeyringController= require('@getsafle/vault-bitcoin-controller').KeyringController ;
+const TezosKeyringController= require('@getsafle/vault-tezos-controller').KeyringController ;
 const bip39 = require('bip39');
 
 const helper = require('../utils/helper');
@@ -51,8 +52,11 @@ class Vault extends Keyring {
     }
 
     initializeSupportedChainKeyringController(mnemonic) {
-        const keyringController = new BitcoinKeyringController({mnemonic:mnemonic});
-        this["bitcoin"] = keyringController;
+        const bitcoinKeyringController = new BitcoinKeyringController({mnemonic:mnemonic});
+        this["bitcoin"] = bitcoinKeyringController;
+
+        const tezosKeyringController = new TezosKeyringController({mnemonic:mnemonic});
+        this["tezos"] = tezosKeyringController;
     }
 
     async generateMnemonic(entropy) {
@@ -146,7 +150,7 @@ class Vault extends Keyring {
 
             const keyringInstance = await helper.getCoinInstance(chainData.toLowerCase(), mnemonic);
 
-            const accArray = await helper.getAccountsFromLogs(keyringInstance, vaultState, recoverMechanism, logs);
+            const accArray = await helper.getAccountsFromLogs(keyringInstance, vaultState, recoverMechanism, logs, chainData);
             const numberOfAcc = accArray.length;
 
             rawVault[chainData.toLowerCase()] = { public: accArray, numberOfAcc }